stdint: correct INT64_MIN and INT64_MAX definitions
Numbers written in hex with suffix LL may be treated
as unsigned long long int by compilator!!!

JIRA: RTOS-219

+++ include/stdint.h
@@ -35,12 +35,12 @@
 #define INT8_MIN  ((int8_t)0x80)
 #define INT16_MIN ((int16_t)0x8000)
 #define INT32_MIN ((int32_t)0x80000000)
-#define INT64_MIN 0x8000000000000000LL
+#define INT64_MIN ((int64_t)0x8000000000000000LL)
 
 #define INT8_MAX  ((int8_t)0x7f)
 #define INT16_MAX ((int16_t)0x7fff)
 #define INT32_MAX ((int32_t)0x7fffffff)
-#define INT64_MAX 0x7fffffffffffffffLL
+#define INT64_MAX ((int64_t)0x7fffffffffffffffLL)
 
 #define UINT8_MAX  0xff
 #define UINT16_MAX 0xffff
